+++ assets/src/edit-story/elements/image/test/output.js
@@ -39,6 +39,6 @@ describe('Image output', () => {
       box: { width: 1080, height: 1920, x: 50, y: 100, rotationAngle: 0 },
     };
 
-    await expect(<ImageOutput {...props} />).toBeValidAMPStoryElement();
+    await expect(<ImageOutput {...props} />).toBeValidAMP();
   });
 });
