+++ assets/src/edit-story/elements/text/test/output.js
@@ -35,6 +35,6 @@ describe('Text output', () => {
       box: { width: 1080, height: 1920, x: 50, y: 100, rotationAngle: 0 },
     };
 
-    await expect(<TextOutput {...props} />).toBeValidAMPStoryElement();
+    await expect(<TextOutput {...props} />).toBeValidAMP();
   });
 });
