+++ package.json
@@ -43,6 +43,7 @@
     "uuid": "^7.0.2"
   },
   "devDependencies": {
+    "@ampproject/toolbox-optimizer": "^2.0.0",
     "@babel/cli": "^7.8.4",
     "@babel/core": "^7.8.7",
     "@babel/plugin-proposal-class-properties": "^7.8.3",
